WinHTTP WinINet support more than 4GB size file

+++ wget/stdafx.h
@@ -50,7 +50,7 @@ struct ProgressInvoke {
 };
 
 struct ProgressCallback {
-	void(*impl)(size_t rate, void *userdata);
+	void(*impl)(uint64_t rate, void *userdata);
 	void *userdata;
 };
